Fix the build warning
drivers/hwtracing/intel_th/gth.c:572:19: warning: unused variable ‘th’ [-Wunused-variable]
struct intel_th *th = to_intel_th(thdev);

drivers/platform/x86/intel_pstore_pram.c:97:2: warning: format ‘%lx’ expects argument of type ‘long unsigned int’, but argument 2 has type ‘phys_addr_t’ [-Wformat=]
   pr_info("registered pram device, addr=0x%lx, size=0x%lx\n",

drivers/staging/android/sbl/sblbc.c:119:3: warning: format ‘%X’ expects argument of type ‘unsigned int’, but argument 2 has type ‘size_t’ [-Wformat=]
    pr_err("Kernel Addr=0x%X, data=0x%X\n", NVRAM_START_ADDRESS + offset + i, *(unsigned char *)(data + i));

drivers/staging/android/sbl/sblbc.c:314:1: warning: label ‘done’ defined but not used [-Wunused-label]
  done:

+++ drivers/platform/x86/intel_pstore_pram.c
@@ -95,7 +95,7 @@ static int register_pram_dev(unsigned long mem_address,
 	}
 
 	pr_info("registered pram device, addr=0x%lx, size=0x%lx\n",
-		pram_data->mem_address, pram_data->mem_size);
+		(unsigned long)pram_data->mem_address, (unsigned long)pram_data->mem_size);
 
 	return 0;
 }
